cleaned the html
I saw a bunch of tutorials to make my html as clean as possible. I don't have much experience before and most of the thigsn i know are self taught. Eventually I realised that my Html code was a mess so I tried to make it cleaner as well as added some projects from the programming bootcamp.